+++ scripts/dns-test.js
@@ -17,22 +17,26 @@ const createSocket = (port) => new Promise((resolve, reject) => {
 
 const domainRegex = /\s([a-z0-9\-]+\.privacytests3\.org)[\s.]/g;
 
+let isObserving = false;
+
 const observeDomains = async () => {
+  if (isObserving) {
+    // We are already observing.
+    return;
+  }
   const socket = await createSocket(9999);
   socket.on('data', (data) => {
     const text = data.toString();
     const matches = [...text.matchAll(domainRegex)];
     for (const match of matches) {
       const domain = match[1];
-      domainsSeen.add(domain);
-      console.log(`${Date.now()}-- observed domain: ${domain}`);
+      if (!domainsSeen.has(domain)) {
+        domainsSeen.add(domain);
+        console.log(`${Date.now()}-- observed domain: ${domain}`);
+      }
     }
   });
-  const stop = () => {
-    socket.destroy();
-    domainsSeen.clear();
-  };
-  return stop;
+  isObserving = true;
 };
 
 const hasSeenDomain = (domain) => {
@@ -62,6 +66,9 @@ const setCountry = (browserSessions, countryCode) =>
     if (browser._defaults.basedOn === 'firefox') {
       await browser.setPref('doh-rollout.home-region', countryCode.toUpperCase());
     }
+    if (browser._defaults.basedOn === 'chromium') {
+      // Set kCFLocaleCountryCode somehow
+    }
   }));
 
 // execAsync(`mullvad relay set location ${countryCode}`);
@@ -89,12 +96,17 @@ const locationDefinition = (countryName, countryCode) =>
   });
 
 const dnsTestDefinitions = [
-  ispDefinition('Comcast', '75.75.75.75'),
-  ispDefinition('Comodo', '8.26.56.26'),
-  ispDefinition('Cox', '68.105.28.11'),
+  //ispDefinition('Comcast', '75.75.75.75'),
+  //ispDefinition('Spectrum (Charter)', '209.18.47.61'),
+  //ispDefinition('AT&T', '68.94.156.1'),
+  //ispDefinition('Verizon', '8.238.64.14'),
+  //ispDefinition('Cox', '68.105.28.11'),
+  //ispDefinition('Orange', '80.10.246.2'),
+  //ispDefinition('BT', '62.6.40.178'),
   ispDefinition('Cloudflare', '1.1.1.1'),
   ispDefinition('Google', '8.8.8.8'),
   ispDefinition('Quad9', '9.9.9.9'),
+  ispDefinition('Comodo', '8.26.56.26'),
   locationDefinition('Brazil', 'br'),
   locationDefinition('China', 'cn'),
   locationDefinition('Germany', 'de'),
@@ -105,7 +117,7 @@ const dnsTestDefinitions = [
 ];
 
 const checkForSecureDns = async (browserSession) => {
-  await sleepMs(60000);
+  await sleepMs(20000);  // Wait enough time for browsers to enable DoH.
   const testDomain = generateRandomTestDomain();
   console.log(`${Date.now()} -- opening ${testDomain}`);
   browserSession.browser.openUrl(`http://${testDomain}/`);
@@ -118,18 +130,19 @@ const testIfDnsIsEncrypted = async (browserSessions, { ip, country }) => {
   await Promise.all(browserSessions.map(session => session.browser.kill()));
   const preferredNetworkService = systemNetworkSettings.getPreferredNetworkService();
   await systemNetworkSettings.setDNS(preferredNetworkService, ip ?? '162.243.184.122');
-  await setCountry(browserSessions, country ?? 'us');
+  await setCountry(browserSessions, country ?? 'aq'); // Antarctica by default
   await sleepMs(2000);
   await Promise.all(browserSessions.map(session => session.browser.launch(false)));
   return await Promise.all(browserSessions.map(checkForSecureDns));
 };
 
 const runDnsTests = async (browserSessions) => {
-  const stopObserving = await observeDomains();
+  await observeDomains();
   const preferredNetworkService = systemNetworkSettings.getPreferredNetworkService();
   const originalDnsIps = systemNetworkSettings.getDNS(preferredNetworkService);
   // Restart all browsers with a fresh profile:
   await Promise.all(browserSessions.map(session => session.browser.restart(true)));
+  await sleepMs(4000);
   const results = [];
   for (const testDef of dnsTestDefinitions) {
     const passedList = await testIfDnsIsEncrypted(browserSessions, testDef);
@@ -140,8 +153,7 @@ const runDnsTests = async (browserSessions) => {
       results[i].dns[testDef.name] = { passed, description: testDef.description, 'leak detected': !passed };
     }
   }
-  systemNetworkSettings.setDNS(preferredNetworkService, originalDnsIps);
-  stopObserving();
+  systemNetworkSettings.setDNS(preferredNetworkService, []);
   return results;
 };